Implemented filter pipeline
You can now add task-specific filters. As a test I've included `AppleFilter`, which filters out any example which contains the substring "apple". Absolutely no one will use it but I think it's funny to put in the repo and so there it will be.
Real, actual usable filters coming soon.